feat: sparo-ci sparse to sparo-ci checkout
